pyproject.toml
[tool.poetry]
name = "fastapi_app"
version = "0.1.0"
description = ""
authors = ["olegd96 <chepalin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-jose = "3.3.0"
bcrypt = "4.1.1"
aiofiles = "23.2.1"
aiohttp = "3.9.1"
aioshutil = "1.3"
aiosignal = "1.3.1"
alembic = "1.13.0"
amqp = "5.2.0"
annotated-types = "0.6.0"
anyio = "3.7.1"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
attrs = "23.1.0"
autoflake = "2.2.1"
billiard = "4.2.0"
black = "23.12.1"
celery = "5.3.6"
certifi = "2023.11.17"
cffi = "1.16.0"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
cryptography = "41.0.7"
distlib = "0.3.7"
dnspython = "2.4.2"
ecdsa = "0.18.0"
email-validator = "2.1.0.post1"
fastapi = "0.104.1"
fastapi-cache2 = "0.2.1"
fastui = "0.2.0"
filelock = "3.13.1"
flake8 = "7.0.0"
flower = "2.0.1"
frozenlist = "1.4.1"
greenlet = "3.0.1"
gunicorn = "21.2.0"
h11 = "0.14.0"
httpcore = "1.0.2"
httptools = "0.6.1"
httpx = "0.25.2"
humanize = "4.9.0"
idna = "3.6"
iniconfig = "2.0.0"
isort = "5.13.2"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
kombu = "5.3.4"
mako = "1.3.0"
markupsafe = "2.1.3"
mccabe = "0.7.0"
multidict = "6.0.4"
mypy = "1.7.1"
mypy-extensions = "1.0.0"
nodeenv = "1.8.0"
orjson = "3.9.10"
packaging = "23.2"
passlib = "1.7.4"
pathspec = "0.12.1"
pendulum = "3.0.0"
pillow = "10.1.0"
platformdirs = "4.1.0"
pluggy = "1.3.0"
prometheus-client = "0.19.0"
prompt-toolkit = "3.0.43"
pyasn1 = "0.5.1"
pycodestyle = "2.11.1"
pycparser = "2.21"
pycryptodome = "3.19.0"
pydantic = "2.5.2"
pydantic-extra-types = "2.1.0"
pydantic-settings = "2.1.0"
pydantic-core = "2.14.5"
pyflakes = "3.2.0"
pyright = "1.1.344"
pytest = "7.4.4"
pytest-asyncio = "0.21.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-json-logger = "2.0.7"
python-multipart = "0.0.6"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
redis = "4.6.0"
rsa = "4.9"
sentry-sdk = "1.39.2"
six = "1.16.0"
sniffio = "1.3.0"
sqladmin = "0.16.0"
sqlalchemy = "2.0.23"
starlette = "0.27.0"
time-machine = "2.13.0"
tornado = "6.4"
types-aiofiles = "23.2.0.0"
types-passlib = "1.7.7.13"
types-pyasn1 = "0.5.0.1"
types-python-jose = "3.3.4.8"
typing-extensions = "4.9.0"
tzdata = "2023.3"
ujson = "5.8.0"
urllib3 = "2.1.0"
uvicorn = "0.24.0.post1"
vine = "5.1.0"
virtualenv = "20.25.0"
watchfiles = "0.21.0"
wcwidth = "0.2.12"
websockets = "12.0"
wtforms = "3.1.1"
yarl = "1.9.4"
aiobotocore = "2.12.3"
fastapi-cli = "^0.0.4"
aio-pika = "^9.4.2"
motor = "^3.5.1"
fastapi-versioning = "^0.10.0"
prometheus-fastapi-instrumentator = "6.1.0"
pytest-cov = "^5.0.0"
grpcio-tools = "^1.66.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.autoflake]
check_diff = true
imports = ["sqlalchemy", "app"]

[tool.pyright]
include = ["app"]

